Refactor main builder class
