+++ scribe/constants.py
@@ -1,0 +1,2 @@
+# Amount of notes shown on one page
+SINGLE_PAGE_NOTE_LIMIT = 5
